Reorganised files and cleaned up code
