+++ Assets/Scripts/Movement/ICharacterMovementSubscriber.cs
@@ -1,0 +1,7 @@
+namespace Movement
+{
+    internal interface ICharacterMovementSubscriber
+    {
+        public void HandleCharacterMovement();
+    }
+}
