+++ Assets/Scripts/Movement/ICharacterMovementSubscriber.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 878f924e33e34a6e9ecf2689efa31d10
+timeCreated: 1693860298
