+++ Assets/Scripts/Movement/PlayerMovementObserver.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 20f70d98f47b41ca9ff8190ab9feb535
+timeCreated: 1693860266
